move pfs cmds tests off kubernetes
